+++ packages/core/src/hook/hook-collector.ts
@@ -1,6 +1,6 @@
 import type { TSESTreeFunction } from "@eslint-react/ast";
 import { getFunctionIdentifier } from "@eslint-react/ast";
-import { O } from "@eslint-react/tools";
+import { F, O } from "@eslint-react/tools";
 import type { ESLintUtils, TSESTree } from "@typescript-eslint/utils";
 import ShortUniqueId from "short-unique-id";
 
@@ -12,29 +12,30 @@ const uid = new ShortUniqueId({ length: 10 });
 
 export function useHookCollector() {
   const hooks = new Map<string, ERHook>();
-  const functionStack: TSESTreeFunction[] = [];
+  const fStack: [node: TSESTreeFunction, id: O.Option<string>][] = [];
   const onFunctionEnter = (node: TSESTreeFunction) => {
-    functionStack.push(node);
-    const currentFn = functionStack.at(-1);
-    if (!currentFn) return;
-    const id = getFunctionIdentifier(currentFn);
+    const id = getFunctionIdentifier(node);
     const name = O.flatMapNullable(id, (id) => id.name);
-    if (O.isSome(id) && O.isSome(name) && isReactHookName(name.value)) {
-      const key = uid.rnd();
-      hooks.set(key, {
-        _: key,
-        id,
-        kind: "function",
-        name,
-        node: currentFn,
-        flag: 0n,
-        hint: 0n,
-        hookCalls: [],
-      });
+    const isHook = O.isSome(id) && O.isSome(name) && isReactHookName(name.value);
+    if (!isHook) {
+      fStack.push([node, O.none()]);
+      return;
     }
+    const key = uid.rnd();
+    fStack.push([node, O.some(key)]);
+    hooks.set(key, {
+      _: key,
+      id,
+      kind: "function",
+      name,
+      node,
+      flag: 0n,
+      hint: 0n,
+      hookCalls: [],
+    });
   };
   const onFunctionExit = () => {
-    functionStack.pop();
+    fStack.pop();
   };
   const ctx = {
     getAllHooks(_: TSESTree.Program): typeof hooks {
@@ -48,26 +49,23 @@ export function useHookCollector() {
     ":function[type]": onFunctionEnter,
     ":function[type]:exit": onFunctionExit,
     "CallExpression[type]"(node) {
-      const currentFn = functionStack.at(-1);
-      if (!currentFn) return;
-      // Detect the number of other hooks called inside the current hook
-      // Hooks that are not call other hooks are redundant
-      // In the realm of React, hooks are like colored functions, and defining a custom hook that doesn't call other hooks is like defining a generator function that doesn't yield or an async function that doesn't await.
-      // "Custom Hooks may call other Hooks (that’s their whole purpose)." from https://react.dev/warnings/invalid-hook-call-warning
-      // Further Reading: https://react.dev/learn/reusing-logic-with-custom-hooks#should-all-functions-called-during-rendering-start-with-the-use-prefix
-      if (isReactHookCall(node)) {
-        const hook = Array
-          .from(hooks.values())
-          .find((hook) => hook.node === currentFn);
-        if (!hook) return;
-        hooks.set(hook._, {
-          ...hook,
-          hookCalls: [
-            ...hook.hookCalls,
-            node,
-          ],
-        });
-      }
+      if (!isReactHookCall(node)) return;
+      const [fNode, hookId] = fStack.at(-1) ?? [];
+      if (!fNode || !hookId) return;
+      F.pipe(
+        O.Do,
+        O.bind("id", () => hookId),
+        O.bind("hook", ({ id }) => O.fromNullable(hooks.get(id))),
+        O.map(({ id, hook }) => {
+          hooks.set(id, {
+            ...hook,
+            hookCalls: [
+              ...hook.hookCalls,
+              node,
+            ],
+          });
+        }),
+      );
     },
   } as const satisfies ESLintUtils.RuleListener;
   return { ctx, listeners } as const;