+++ packages/plugins/eslint-plugin-react-hooks-extra/src/rules/no-redundant-custom-hook.ts
@@ -1,0 +1,48 @@
+import { useHookCollector } from "@eslint-react/core";
+import type { CamelCase } from "string-ts";
+
+import { createRule } from "../utils";
+
+export const RULE_NAME = "no-redundant-custom-hook";
+
+export type MessageID = CamelCase<typeof RULE_NAME>;
+
+// Detect the number of other hooks called inside the current hook
+// Hooks that are not call other hooks are redundant
+// In React, hooks are like colored functions, and defining a custom hook that doesn't call other hooks is like defining a generator function that doesn't yield or an async function that doesn't await.
+// "Custom Hooks may call other Hooks (that’s their whole purpose)." from https://react.dev/warnings/invalid-hook-call-warning
+// Further Reading: https://react.dev/learn/reusing-logic-with-custom-hooks#should-all-functions-called-during-rendering-start-with-the-use-prefix
+export default createRule<[], MessageID>({
+  meta: {
+    type: "problem",
+    docs: {
+      description: "enforce custom hooks to use at least one other hook inside",
+    },
+    messages: {
+      noRedundantCustomHook: "A custom hook '{{name}}' should use at least one other hook.",
+    },
+    schema: [],
+  },
+  name: RULE_NAME,
+  create(context) {
+    const { ctx, listeners } = useHookCollector();
+
+    return {
+      ...listeners,
+      "Program:exit"(node) {
+        const allHooks = ctx.getAllHooks(node);
+        for (const { name, node, hookCalls } of allHooks.values()) {
+          if (hookCalls.length > 0) continue;
+          context.report({
+            messageId: "noRedundantCustomHook",
+            node,
+            data: {
+              name: name.value,
+            },
+          });
+        }
+      },
+    };
+  },
+  defaultOptions: [],
+});